+++ integration-tests/gatsby-source-wordpress/test-fns/plugin-options.js
@@ -1,6 +1,4 @@
-const {
-  default: fetchGraphql,
-} = require("gatsby-source-wordpress/dist/utils/fetch-graphql")
+const { fetchGraphql } = require("./test-utils/graphql")
 
 describe(`plugin options`, () => {
   test(`Type.exclude option removes types from the schema`, async () => {