+++ packages/gatsby-source-wordpress/src/models/index.ts
@@ -6,8 +6,30 @@ import wpHooks from "./wp-hooks"
 import previewStore from "./preview"
 import develop from "./develop"
 import postBuildWarningCounts from "./post-build-warning-logs"
+import { Models, createModel } from "@rematch/core"
 
-export default {
+export const settings = createModel<IRootModel>()({
+  state: 0,
+  reducers: {
+    increment: (state, payload: number) => state + payload,
+  },
+  effects: () => {
+    return {}
+  },
+})
+
+export interface IRootModel extends Models<IRootModel> {
+  remoteSchema: typeof remoteSchema
+  gatsbyApi: typeof gatsbyApi
+  logger: typeof logger
+  imageNodes: typeof imageNodes
+  wpHooks: typeof wpHooks
+  previewStore: typeof previewStore
+  develop: typeof develop
+  postBuildWarningCounts: typeof postBuildWarningCounts
+}
+
+const models: IRootModel = {
   remoteSchema,
   gatsbyApi,
   logger,
@@ -17,3 +39,5 @@ export default {
   develop,
   postBuildWarningCounts,
 }
+
+export default models