+++ packages/gatsby-source-wordpress/src/steps/create-schema-customization/transform-fields/transform-union.js
@@ -1,9 +1,10 @@
 import { buildTypeName } from "~/steps/create-schema-customization/helpers"
-import { findNamedTypeName, introspectionFieldTypeToSDL } from "../helpers"
+import { introspectionFieldTypeToSDL } from "../helpers"
 
-export const transformUnion = ({ field, fieldName }) => {
+export const transformUnion = ({ field, fieldName, pluginOptions }) => {
+  const prefix = pluginOptions.schema.typePrefix
   return {
-    type: buildTypeName(field.type.name),
+    type: buildTypeName(field.type.name, prefix),
     resolve: (source, _, context) => {
       const resolvedField =
         source[fieldName] ||
@@ -25,7 +26,8 @@ export const transformUnion = ({ field, fieldName }) => {
   }
 }
 
-export const transformListOfUnions = ({ field, fieldName }) => {
+export const transformListOfUnions = ({ field, fieldName, pluginOptions }) => {
+  const prefix = pluginOptions.schema.typePrefix
   const typeSDLString = introspectionFieldTypeToSDL(field.type)
 
   return {
@@ -46,7 +48,7 @@ export const transformListOfUnions = ({ field, fieldName }) => {
         const node = item?.id
           ? context.nodeModel.getNodeById({
               id: item.id,
-              type: buildTypeName(item.__typename),
+              type: buildTypeName(item.__typename, prefix),
             })
           : null