+++ packages/gatsby-source-wordpress/src/steps/ingest-remote-schema/build-queries-from-introspection/recursively-transform-fields.js
@@ -1,4 +1,4 @@
-import store from "~/store"
+import { getStore } from "~/store"
 import {
   getTypeSettingsByType,
   findNamedTypeName,
@@ -26,7 +26,7 @@ export const transformInlineFragments = ({
   buildingFragment = false,
   ancestorTypeNames: parentAncestorTypeNames = [],
 }) => {
-  const state = store.getState()
+  const state = getStore().getState()
 
   if (!typeMap) {
     typeMap = state.remoteSchema.typeMap
@@ -76,7 +76,7 @@ export const transformInlineFragments = ({
       possibleType.type = { ...type }
 
       // save this type so we can use it in schema customization
-      store.dispatch.remoteSchema.addFetchedType(type)
+      getStore().dispatch.remoteSchema.addFetchedType(type)
 
       const isAGatsbyNode = gatsbyNodesInfo.typeNames.includes(
         possibleType.name
@@ -566,7 +566,7 @@ const transformFields = ({
 
       if (transformedField) {
         // save this type so we know to use it in schema customization
-        store.dispatch.remoteSchema.addFetchedType(field.type)
+        getStore().dispatch.remoteSchema.addFetchedType(field.type)
       }
 
       const typeName = findNamedTypeName(field.type)
@@ -664,7 +664,7 @@ const recursivelyTransformFields = ({
   const {
     gatsbyApi: { pluginOptions },
     remoteSchema: { fieldBlacklist, fieldAliases, typeMap, gatsbyNodesInfo },
-  } = store.getState()
+  } = getStore().getState()
 
   const {
     schema: { queryDepth, circularQueryLimit },