+++ packages/gatsby-source-wordpress/src/steps/source-nodes/helpers.js
@@ -1,9 +1,9 @@
-import store from "~/store"
+import { getStore } from "~/store"
 import { camelCase } from "lodash"
 import { findNamedTypeName } from "../create-schema-customization/helpers"
 
 export const getTypeInfoBySingleName = singleName => {
-  const { typeMap } = store.getState().remoteSchema
+  const { typeMap } = getStore().getState().remoteSchema
 
   const rootField = typeMap
     .get(`RootQuery`)
@@ -17,7 +17,7 @@ export const getTypeInfoBySingleName = singleName => {
 }
 
 export const getQueryInfoBySingleFieldName = singleName => {
-  const { nodeQueries } = store.getState().remoteSchema
+  const { nodeQueries } = getStore().getState().remoteSchema
   const queryInfo = Object.values(nodeQueries).find(
     q =>
       q.typeInfo.singularName === singleName ||
@@ -28,7 +28,7 @@ export const getQueryInfoBySingleFieldName = singleName => {
 }
 
 export const getQueryInfoByTypeName = typeName => {
-  const { nodeQueries } = store.getState().remoteSchema
+  const { nodeQueries } = getStore().getState().remoteSchema
 
   const queryInfo = Object.values(nodeQueries).find(
     q => q.typeInfo.nodesTypeName === typeName