+++ packages/gatsby-source-wordpress/src/steps/source-nodes/update-nodes/fetch-node-updates.js
@@ -4,6 +4,7 @@ import { formatLogMessage } from "~/utils/format-log-message"
 import { getGatsbyApi } from "~/utils/get-gatsby-api"
 import { getPersistentCache } from "~/utils/cache"
 import { needToTouchNodes } from "../../../utils/gatsby-features"
+import { withPluginKey } from "~/store"
 
 export const touchValidNodes = async () => {
   if (!needToTouchNodes) {
@@ -42,7 +43,7 @@ const fetchAndApplyNodeUpdates = async ({
   activity.start()
 
   if (!since) {
-    since = await cache.get(LAST_COMPLETED_SOURCE_TIME)
+    since = await cache.get(withPluginKey(LAST_COMPLETED_SOURCE_TIME))
   }
 
   // Check with WPGQL to create, delete, or update cached WP nodes